# src/api-docs/openapi-specs.yml
openapi: 3.0.0
info:
  title: PowerX Backend Todo
  description: Capstone Project for PowerX Backend Module
  version: 1.0.0
servers:
  - url: https://grumbeard-powerx-todo.herokuapp.com
    description: Production server
  - url: /
    description: Development server (local)
components:
  schemas:
...
    Item:
      type: object
      required:
        - id
        - description
        - todo_list_id
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        description:
          type: string
          minLength: 1
        todo_list_id:
          type: integer
          format: int64
          minimum: 1
        is_completed:
          type: boolean
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
paths:
...
  /item:
    post:
      tags:
        - item
      summary: Create an Item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - description
                - todo_list_id
              properties:
                description:
                  type: string
                  minLength: 1
                todo_list_id:
                  type: integer
                  format: int64
                  minimum: 1
                is_completed:
                  type: boolean
            examples:
              item-details:
                summary: Example of Item details
                value:
                  description: "buy bread"
                  todo_list_id: 1
                  is_completed: false
      responses:
        '200':
          description: Created Item returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
              examples:
                item:
                  summary: Example of Item
                  value:
                    id: 1
                    description: "pick up sheep"
                    todo_list_id: 1
                    is_completed: false
        '400':
          description: Invalid Item details provided
        '403':
          description: Unauthorized to create Item for this TodoList
        default:
          description: Unexpected error  
  /item/{id}:
    patch:
      tags:
        - item
      summary: Update an Item by Id
      parameters:
        - name: id
          in: path
          required: true
          description: Item ID
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  minLength: 1
                is_completed:
                  type: boolean
            examples:
                item_details:
                  summary: Example of Item details
                  value:
                    description: "put down sheep"
      responses:
        '200':
          description: Updated Item returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
              examples:
                item:
                  summary: Example of Item
                  value:
                    id: 1
                    description: "pick up sheep"
                    todo_list_id: 1
        '400':
          description: Item does not exist
        '403':
          description: Unauthorized to update Item
        default:
          description: Unexpected error
    delete:
      tags:
        - item
      summary: Delete an Item by Id
      parameters:
        - name: id
          in: path
          required: true
          description: Item ID
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Deleted (Soft-Delete) Item returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
              examples:
                item:
                  summary: Example of Item
                  value:
                    id: 1
                    description: "pick up sheep"
                    todo_list_id: 1
        '400':
          description: Item does not exist
        '403':
          description: Unauthorized to delete Item
        default:
          description: Unexpected error